crucible-mc: update treatment of assumptions
Before this commit, we were not handling assumptions.  The treatment of
assumptions requires some care: we don't get to describe the assumptions of
a given block within the transition for this block.

Instead, we must make sure the assumptions hold at every incoming transition
into the block.  As such, each transition must be conditioned by the
assumptions of its target block.  This ought to invalidate all the traces
that do not validate the assumptions of the block they transfer control to.