Fix issues with the Crux HTML UI
Make sure all the paths that are used in the source view pane and the
verification conditions list are canonicalized in the same way.  This
helps to ensure the javascript doesn't get confused when searching for
files by name.

In addition, pipe in a bit more error information into the UI, using
new `details-short` and `details-long` fields of the side condition
objects.  I've displayed this information in a new `error-pane` div
above the source code view.  Someone with better UI skills than I could
perhaps to a better job finding a way to display this information.

Fixes #534, #517